Fix #93 highlighting of complex parencites calls
Note that a repeated capturing group will only capture the last
iteration. Put a capturing group around the repeated group to capture
all iterations.

To capture all iterations of the repeated group, you can enclose the
entire repeated group in another capturing group. However, this approach
captures the entire sequence as a single string, not individual matches.

We use patterns and include to solve this issue.